Add support for Single-Quoted string to hic
Recently, there was a discussion about escaping non-translated strings
in the HID-Language. Resulting in the knowledge that there currently is
no way to add spaces, colons and other special characters to a
non-translated string in HID.

While I was having the lexer code on screen for an unrelated feature, I
just doubled the lexer part that takes care of parsing double quoted
strings (for translation).

With this change, hic can compile strings like this:

  Action FooDance {
    Text            "Dance the &Foo";
    _ConnectContext 'GlobalFoo::self()->getTheFooDancer()';
    _ConnectTo      'onDanceTheFoo()';
  };

However, identifiers are converted to C++ and must produce valid C++.
That means:

  Action 'Foo Dance';

would produce C++ like:

  QAction* actFoo Dance;

which is - of course - a syntactical failure...