Change few QLists with QVectors
Actually, QList usage of libHeaven is quite distinctive. Here we're
using QList mostly as a list of pointer to something (Which is deemed to
be good usage of QList). This commit changes the few places where we
used QList< struct > with QVector< struct > and leaves one place of such
intact, as it's easier to use the QList API in that special place.