Guard against double deletion
Some weeks agao, I made containers delete theirselves when their last
child got missing in action.
If this happens during a mode switch (i.e. on application termination or
closing of last repo in mgv), the temporary lists of containers to
delete is not (and cannot be) updated. Thus we need to guard that list
against double deletion.