Allow to build against build-dir, too
The hic.cmake file isn't exactly present where we're looking for it.
This makes it available in the install-dir in the same directory as
in the source dir and configures a wrapper in the build dir, that
will look into the source dir.

This is necessary at least on Windows.